Fixed Loss of Fraction problem with CalculateDensity Method in Graph Class

+++ Oz.Algorithms/Math/Graph/Graph.cs
@@ -31,7 +31,7 @@ namespace Oz.Algorithms.Math.Graph
 	        {
 		        CalculateMaximumVertex();
 	        }
-            Density = MaximumVertexCount / Nodes.Count();
+            Density = (double)MaximumVertexCount / (double)Nodes.Count();
         }
 
         /// <summary>
